All tests passing except 1
Lot's of fixes in here. Made finding Pallets by index not be positional in an array, now it's using a properly indexed Map.
Add TransactionPayment Pallet and it's assocaited Event because that event was being sent in the BalancesTest

Any test that is disabled should have comments but I will call them "event tests" are disabled because it requires a behavior in the test container that Frequency doesn't currently support. I tested these against Vanilla Substrate (whcih can create blocks passively every 3 seconds) and they actually pass, so the library does work and that's all I care about.